+++ package.json
@@ -16,6 +16,7 @@
     "react-router": "^4.2.0",
     "react-router-dom": "^4.2.2",
     "react-scripts": "1.0.17",
+    "recompose": "^0.30.0",
     "styled-components": "^3.2.1"
   },
   "scripts": {
