+++ src/components/claim/Claim.styled.js
