+++ src/components/skill-card/SkillCard.styled.js
