+++ src/pages/curriculum-vitae/job-card/JobCard.js
