+++ src/pages/curriculum-vitae/job-card/JobCard.styled.js
