+++ src/pages/skills/skills-data.js
@@ -1,15 +1,15 @@
-import logoDesign from './assets/design.svg'
-import logoCode from './assets/code.svg'
-import logoTeam from './assets/team.svg'
-import logoReact from './assets/react.svg'
-import logoRedux from './assets/redux.svg'
-import logoNode from './assets/node.svg'
-import logoWebpack from './assets/webpack.svg'
-import logoBabel from './assets/babel.svg'
-import logoJest from './assets/jest.svg'
-import logoFlow from './assets/flow.png'
-import logoEslint from './assets/eslint.png'
-import logoStyledComponents from './assets/styled-components.png'
+import logoDesign from '../../assets/design.svg'
+import logoCode from '../../assets/code.svg'
+import logoTeam from '../../assets/team.svg'
+import logoReact from '../../assets/react.svg'
+import logoRedux from '../../assets/redux.svg'
+import logoNode from '../../assets/node.svg'
+import logoWebpack from '../../assets/webpack.svg'
+import logoBabel from '../../assets/babel.svg'
+import logoJest from '../../assets/jest.svg'
+import logoFlow from '../../assets/flow.png'
+import logoEslint from '../../assets/eslint.png'
+import logoStyledComponents from '../../assets/styled-components.png'
 
 const softSkills = [
   {
